opencsd: etmv4: Rework of etmv4 decode class to simplify.
Multiple re-entrant paths were possible through this code, making
maintenance and modification difficult.

An out packet buffer has been added to simplify this code, so that
all out packets for a single input packet are collated before being
sent to the output interface.

This will make the implementation of speculative trace much easier
as the code is far clearer and more linear in the trace element stack
area.

Signed-off-by: Mike Leach <mike.leach@linaro.org>